\encoding{latin1}
\name{CATFLOW-package}
\alias{Catflow-package}
\alias{CATFLOW}
\alias{catflow}
\alias{Catflow}
\docType{package}
\title{
CATFLOW Utilities
}

\description{
Utilities for pre- and postprocessing for the hydrological model CATFLOW
}

\details{
\tabular{ll}{
Package: \tab Catflow\cr
Type: \tab Package\cr
Version: \tab 0.984\cr
Date: \tab 2012-03-29\cr
License: \tab GPL-2\cr
LazyLoad: \tab yes\cr
}
}

\author{
Jan \enc{Wienhöfer}{Wienhoefer}  <j.wienhoefer@gmx.de>
}
