\name{make.geometry}

\alias{make.geometry}

\title{
Generate geometry for CATFLOW
}
\description{
This function generates a set of curvilinear orthogonal coordinates and writes the
corresponding geometry file for use with the hydrological model CATFLOW.
}
\usage{
make.geometry(slope.list, 
                    plotting = TRUE, make.output = TRUE,
                     project.path = NULL, 
                      thick.tol = 0.2, ...)
}

\arguments{
  \item{slope.list}{
 a list with the components:
 
    -  xh       : x-coordinate of slope line (northing) [m]
    
    -  yh       : y-coordinate of slope line (easting) [m]
    
    -  zh       : z-coordinate of slope line (surface elevation) [m]
    
    -  tot.area : total area of slope [m²]
    
    -  bh       : slope width [m]
    
    -  dyy      : thickness of profile [m]
    
    -  xsi      : relative discretization in lateral direction
    
    -  eta      : relative discretization in vertical direction
   
  ###  the following parameters may take default values  ###
   
    -  htyp     : type of geometry: (1) constant thickness (default), (2) cake-shape, (3) variable thickness with spline approximation of lower boundary - see Details 
   
    -  numh     : ID for slope (default 1)
      
    -  ho_bez   : point of reference for y-coordinate of slope profile (northing, default 0)
   
    -  re_bez   : point of reference for x-coordinate of slope profile (easting, default 0)
   
    -  z_bez    : point of reference for z-coordinate of slope profile (elevation, default 0)
   
    -  w.aniso  : angle of main direction of anisotropy,  positive counterclockwise (default 0)
   
    -  out.file : name for CATFLOW geofile to produce (no file is generated if empty)
}
  \item{plotting}{
Logical indicating whether plots are to be generated during execution of the function (default T).
}
  \item{make.output}{
Logical indicating whether the specified output file is to be generated by the function (default T).
When no output file name is specified, no file will be generated.
}
# 
  \item{project.path}{
  Path to directory where the specified output file is to be generated (optional) 
}
  \item{thick.tol}{
  Tolerance for thickness when slope type is "3" (default 0.2) - see Details. 
}
  \item{...}{
  Additional parameters, especially graphical parameters for the final plot, e.g. col = 1 for black lines, or 'plottitle' which overrides the default title of the plot. 
}
}
\details{'make.geometry' is primarily used for its side-effect: generating a geometry file
for the CATFLOW model. This file, given by the 'out.file' argument, holds the coordinates of
the simulation nodes and the corresponding metric coefficients calculated with this function.

The slope types (currently three) differ in how the lower boundary is generated. With type
"1", the upper boundary (given as input argument) is shifted downwards in direction of the
tangents at the endpoints by a distance according to the specified thickness 'dyy'.
NOTE: When 'dyy' is a vector with the same length as the slope line ('xh','yh' etc.), each point of 
the upper boundary is shifted downwards according to the corresponding entry in 'dyy'. Thus, an irregular lower 
boundary may be produced.

Type "2" produces a 'cake-shape', that is a horizontal lower boundary at an elevation of
min(zh) - dyy.

When slope type = "3", the lower boundary is generated as a simplification of the top
boundary by shifting a subset of points vertically downwards according to the specified
thickness 'dyy', and interpolating a complete lower boundary with splines. 
This is done in an iterative way by taking more mid-points into account until the final
thickness of the slope layer is within the range (1 +/- thick.tol)*dyy. The thickness
tolerance 'thick.tol' defaults to 0.2, which means that for a desired thickness of 1 m
(dyy = 1), the lower boundary will be iteratively refined until the thickness of the
geometry is between 0.8 and 1.2 m. 
}
\value{
When assigned, the functions returns a list with the components:
 \item{ID}{ID of the slope profile}
 \item{eta, xsi}{Discretization in vertical and lateral direction, respectively}
 \item{hko}{Matrix of local vertical coordinates of simulation nodes [m]}
 \item{sko}{Matrix of local horizontal coordinates of simulation nodes [m]}
 \item{re}{Real-world x-coordinates of simulation nodes (eastings) [m]} 
 \item{ho}{Real-world y-coordinates of simulation nodes (northings) [m]} 
 \item{var.width}{Slope widths for simulation nodes in lateral direction [m]} 
 \item{re_bez, ho_bez, z_bez}{Points of reference for real-world coordinates (x,y,z, respectively)} 
 \item{tot.area}{Surface area of slope profile [m²]}
 \item{width}{Average width of slope profile [m]}   
 \item{length}{Full length of slope profile [m]}   
}

\references{Maurer, T. and E. Zehe (2007) CATFLOW User Guide and Program Documentation (Version Catstat).

The algorithm follows the methodology described in:

Kiefer, E.-M., R. Liedl, G.H. Schmitz and G.J. Seus (1990) Konservative Strömungsmodelle
auf der Basis krummliniger Koordinaten  unter besonderer Berücksichtigung von Wasserbewegungen
im ungesättigt-gesättigten Boden. Berichte der Versuchsanstalt Obernach
und des Lehrstuhls für Wasserbau und Wassermengenwirtschaft der Technischen Universität
München, Nr. 64. ISSN 0939-0308. [in German]
}
\author{
Jan \enc{Wienhöfer}{Wienhoefer}, Erwin Zehe
}
\note{
This R code is an adaption of MATLAB code provided by Erwin Zehe. 
}


\seealso{
\code{\link{read.geofile}}
}
\examples{
  \dontrun{
     simple.slope <- list(
                      xh = seq(1,11, length=20),
                      yh = seq(2,8, length=20),
                      zh = approx(c(8,5),n=20)$y + sin((0:19)/2)/5 ,
                      bh = rep(1,20),
                      tot.area = 12 ,
                      htyp = 1, 
                      dyy = 2,
                      xsi = seq(0,1,length=11),
                      eta = seq(0,1,length= 6),
                      out.file="test.geo"    
                        # other parameters may take default values here
                        # no output file generated when 'out.file' 
                        # is not in list, or when make.output=FALSE
                      )
                      
     test.geom <- make.geometry(simple.slope, make.output=FALSE)                 
  }                    
}


\keyword{ utilities }
