\name{read.evapo.out}
\alias{read.evapo.out}

\title{
Read CATFLOW evapo-transpiration output
}
\description{
Reads a evapo-transpiration output file from a CATFLOW simulation.
}
\usage{
read.evapo.out(filename)
}

\arguments{
  \item{filename}{
Name (and path) of an evapo-transpiration file to read 
}
}

\value{
A data frame with the contents of the evapo-transpiration file.
}
\references{
Maurer, T. and E. Zehe (2007) \emph{CATFLOW User Guide and Program Documentation (Version Catstat)}
}
\author{
Jan \enc{Wienhöfer}{Wienhoefer}
}

\seealso{
\code{\link{read.catf.results}}
}

\examples{
### TO DO ###
## problem: file to read ##
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{utilities}

